+++ drive/__init__.py
@@ -2,7 +2,7 @@
 
 from drive.client import Client
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 
 __all__ = [
     "__version__",