+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Raise a more explicit exception when `upload()` or `upload_file()` is used without an explicit `original_mime_type`
   and libmagic is not installed (#11).
 * Fix the type hint of `parent_id`, which can be a `File` or a `str`
+* Add official support for Python 3.11
 
 ## 0.4.0 (2022/10/12)